requirements/base.txt
django
psycopg2-binary
python-environ
django-cors-headers
django-jazzmin
drf-yasg
djangorestframework
django-filter
djangorestframework-simplejwt==5.2.0
django-allauth==0.51.0
dj-rest-auth==2.2.5
social-auth-app-django==5.0.0